+++ src/index.ts
@@ -36,5 +36,6 @@
 export { LitteraContext, LitteraProvider } from "./LitteraProvider";
 export { default as withLittera } from "./withLittera";
 export { useLittera, useLitteraMethods } from "./hooks";
+export { default as Trans } from './Trans';
 
-export { ILitteraProvider, ITranslations, ITranslated } from "../types/index.d";
+export type { ILitteraProvider, ITranslations, ITranslated } from "../types/index.d";