+++ webpack.config.js
@@ -35,11 +35,6 @@ module.exports = env => {
                 { test: /\.tsx?$/, use: 'awesome-typescript-loader' },
                 { test: /\.js$/, use: 'source-map-loader' },
                 {
-                    test: /\.(js|jsx)$/,
-                    use: 'babel-loader',
-                    exclude: /node_modules/,
-                },
-                {
                     test: /\.css$/,
                     use: ['style-loader', 'css-loader'],
                 },